requirements.txt
# Direct known Python requirements
# (Javascript prerequisites are specified in package.json)

boto3==1.17.21
Django==2.2.24
django-phonenumber-field==5.2.0
django-recaptcha==2.0.6
django-webpack-loader==0.6.0
djangorestframework==3.12.4
# next 8 usually installed through source repo directly.
djaodjin-deployutils==0.6.0
djaodjin-extended-templates==0.2.7
djaodjin-multitier==0.1.21
djaodjin-pages==0.4.1
djaodjin-rules==0.2.8
djaodjin-saas==0.10.1
djaodjin-signup==0.5.1
docutils==0.15.2  # botocore has a constraint: docutils<0.16
jinja2==2.10.1
MarkupSafe==1.1.1  # force installs in case older version is found (CentOS7)
# gunicorn 19.0 is broken (see issue 787)
# gunicorn 19.3.0 has a syntax error in gunicorn/workers/_gaiohttp.py:64
#   but it still seems to work ...
gunicorn==19.7.1
phonenumbers==8.12.6
PyJWT==2.0.1
pytz==2021.1
social-auth-app-django==4.0.0 # v1.2.0 does not support Django>=2.1
whitenoise==5.1.0

# To run with DEBUG=1
django-extensions==2.2.9
django-debug-panel==0.8.3
django-debug-toolbar==2.2.1 # 1.11 does not support Django 2.2

# development
drf-yasg==1.8.0
Faker==3.0.1
flex==6.14.0
pygments==2.2.0
pyYAML==5.1.2
Sphinx==1.8.4
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-websupport==1.1.0
swagger-spec-validator==2.3.0
